Code cleanup
git-svn-id: http://ndddsample.googlecode.com/svn/trunk@50 8496d0c4-1a5a-11de-9308-ab3afd9c0aba

+++ src/app/domain/NDDDSample.Domain/Model/Cargos/Itinerary.cs
@@ -19,7 +19,7 @@
         internal static readonly Itinerary EMPTY_ITINERARY = new Itinerary();
         private static readonly DateTime END_OF_DAYS = DateTime.MaxValue;
         private readonly IList<Leg> legs = new List<Leg>();
-        private long id;
+        private int id;
 
         #region Constr
 
@@ -102,10 +102,10 @@
                 Leg lastLeg = LastLeg;
 
                 if (lastLeg == null)
-                {                  
+                {
                     return END_OF_DAYS;
                 }
-               
+
                 return lastLeg.UnloadTime;
             }
         }
